refactor: Remove CallTitle component and use enum for call type in Call component
