+++ phone-test/src/declarations/call.d.ts
@@ -1,6 +1,6 @@
 interface Call {
   id: string;
-  call_type: CallType;
+  call_type: string;
   created_at: string;
   direction: string;
   from: string;
@@ -14,10 +14,4 @@ interface Call {
 interface Note {
   id: string;
   content: string;
-}
-
-enum CallType {
-  MISSED = 'missed',
-  ANSWERED = 'answered',
-  VOICEMAIL = 'voicemail'
 }
